Initial commit for go package git.ub.uni-freiburg.de/tambora
This package contains all go app's for tambora.org

Initially this will be an app for coding data on tambora. It's located in the coding/
subdirectory. At this point it includes a SLES 11 init file to run it as a server,
a config file, that is read by the init file, a README.md documenting the usage and
a minimal concept tambora-coding.go source file, to test the environment and the
database connection.